Client response refactor Part 1 (#559)
* move client_response into its own file

* changing client_response to api_error as it is more descriptive less confusing

* changing api_error to backend_service_error
