+++ lib/common/client/middleware/response/raise_error.rb
@@ -6,7 +6,7 @@ module Common
         class RaiseError < Faraday::Response::Middleware
           def on_complete(env)
             return if env.success?
-            raise Common::Client::Errors::ClientResponse.new(env.status.to_i, env[:body])
+            raise Common::Client::Errors::BackendServiceError.new(env.status.to_i, env[:body])
           end
         end
       end
