+++ lib/facilities/multi_client.rb
@@ -18,12 +18,12 @@ module Facilities
           Rails.logger.error "GIS request timed out: #{req.url}"
           raise Common::Exceptions::ClientError.new('Facility request timed out', {})
         end
-        raise Common::Client::Errors::ClientResponse.new(req.response.code, {}) unless
+        raise Common::Client::Errors::BackendServiceError.new(req.response.code, {}) unless
           req.response.success?
         result = JSON.parse(req.response.body)
         if result['error']
           Rails.logger.error "GIS returned error: #{result['error']['code']}, message: #{result['error']['message']}"
-          raise Common::Client::Errors::ClientResponse.new(result['error']['code'], {})
+          raise Common::Client::Errors::BackendServiceError.new(result['error']['code'], {})
         end
         result['features']
       end
