+++ spec/controllers/application_controller_spec.rb
@@ -15,7 +15,7 @@ RSpec.describe ApplicationController, type: :controller do
     end
 
     def other_error
-      raise Common::Client::Errors::ClientResponse.new(422, JSON_ERROR)
+      raise Common::Client::Errors::BackendServiceError.new(422, JSON_ERROR)
     end
 
     def client_connection_failed
@@ -49,7 +49,7 @@ RSpec.describe ApplicationController, type: :controller do
     end
   end
 
-  context 'ClientResponseError' do
+  context 'BackendServiceErrorError' do
     subject { JSON.parse(response.body)['errors'].first }
     before(:each) { routes.draw { get 'other_error' => 'anonymous#other_error' } }
     let(:keys_for_production) { %w(title detail code status) }
